+++ frontend/app/components/HeaderClient.tsx
@@ -5,7 +5,7 @@ import {useEffect, useMemo, useState} from 'react'
 import {Orbitron} from 'next/font/google'
 import {urlForImage} from '@/sanity/lib/utils'
 import {MagnifyingGlassIcon} from '@heroicons/react/24/outline'
-import { OriginalSearchModal } from './OriginalSearchModal'
+import SearchModal from '@/app/components/SearchModal'
 
 const orbitron = Orbitron({
   subsets: ['latin'],
@@ -569,7 +569,7 @@ export default function HeaderClient({
       </div>
 
       {/* Search Modal */}
-      <OriginalSearchModal 
+      <SearchModal 
         isOpen={isSearchOpen} 
         onClose={() => setIsSearchOpen(false)} 
         vehicles={timberlineVehicles as any}
